requirements.txt
python-dotenv

numpy
scipy
matplotlib

pypdf
unstructured[md]
html2text
beautifulsoup4
chromadb
wikipedia
numexpr
google-search-results

Flask

openai
tiktoken
langchain
langchain_community
langchain_openai
langchain_chroma
langchain-experimental